add should write to disk checks

+++ source/_addonStore/dataManager.py
@@ -158,7 +158,8 @@ class _DataManager:
 				cacheData = json.load(cacheFile)
 		except Exception:
 			log.exception(f"Invalid add-on store cache")
-			os.remove(cacheFilePath)
+			if NVDAState.shouldWriteToDisk():
+				os.remove(cacheFilePath)
 			return None
 		try:
 			data = cacheData["data"]
@@ -167,7 +168,8 @@ class _DataManager:
 			nvdaAPIVersion = cacheData["nvdaAPIVersion"]
 		except KeyError:
 			log.exception(f"Invalid add-on store cache:\n{cacheData}")
-			os.remove(cacheFilePath)
+			if NVDAState.shouldWriteToDisk():
+				os.remove(cacheFilePath)
 			return None
 		return CachedAddonsModel(
 			cachedAddonData=_createStoreCollectionFromJson(data),
